整理代码

+++ script/archive_mpq.lua
@@ -1,7 +1,50 @@
 local stormlib = require 'ffi.stormlib'
 
+local function get_map_flag(w3i)
+    if not w3i then
+        return 0
+    end
+    return w3i['选项']['关闭预览图']       << 0
+         | w3i['选项']['自定义结盟优先权'] << 1
+         | w3i['选项']['对战地图']        << 2
+         | w3i['选项']['大型地图']        << 3
+         | w3i['选项']['迷雾区域显示地形'] << 4
+         | w3i['选项']['自定义玩家分组']   << 5
+         | w3i['选项']['自定义队伍']       << 6
+         | w3i['选项']['自定义科技树']     << 7
+         | w3i['选项']['自定义技能']       << 8
+         | w3i['选项']['自定义升级']       << 9
+         | w3i['选项']['地图菜单标记']     << 10
+         | w3i['选项']['地形悬崖显示水波'] << 11
+         | w3i['选项']['地形起伏显示水波'] << 12
+         | w3i['选项']['未知1']           << 13
+         | w3i['选项']['未知2']           << 14
+         | w3i['选项']['未知3']           << 15
+         | w3i['选项']['未知4']           << 16
+         | w3i['选项']['未知5']           << 17
+         | w3i['选项']['未知6']           << 18
+         | w3i['选项']['未知7']           << 19
+         | w3i['选项']['未知8']           << 20
+         | w3i['选项']['未知9']           << 21
+end
+
 local mt = {}
 mt.__index = mt
+
+function mt:save(path, w3i, n, encrypt)
+    local hexs = {}
+    hexs[#hexs+1] = ('c4'):pack('HM3W')
+    hexs[#hexs+1] = ('c4'):pack('\0\0\0\0')
+    hexs[#hexs+1] = ('z'):pack(w3i and w3i['地图']['地图名称'] or '未命名地图')
+    hexs[#hexs+1] = ('l'):pack(get_map_flag(w3i))
+    hexs[#hexs+1] = ('l'):pack(w3i and w3i['玩家']['玩家数量'] or 233)
+    io.save(path, table.concat(hexs))
+    self.handle = stormlib.create(path, n, encrypt)
+    if not self.handle then
+        return false
+    end
+    return true
+end
 
 function mt:close()
     return self.handle:close()
@@ -16,6 +59,9 @@ function mt:has_file(name)
 end
 
 function mt:remove_file(name)
+    if self.read then
+        return false
+    end
     return self.handle:remove_file(name)
 end
 
@@ -42,15 +88,13 @@ return function (input, read)
         else
             handle = stormlib.open(input, true)
         end
+        if not handle then
+            return nil
+        end
         if not handle:has_file '(listfile)' then
             message('不支持没有(listfile)的地图')
             return nil
         end
-    else
-        handle = stormlib.open(input)
-    end
-    if not handle then
-        return nil
     end
     return setmetatable({ handle = handle, read = read }, mt)
 end
